+++ app/Http/Controllers/TaskListController.php
@@ -46,6 +46,8 @@ class TaskListController extends Controller
      */
     public function show(TaskList $taskList)
     {
+        abort_unless($taskList->user_id === Auth::id(), 403);
+
         return $this->success([
             'task_list' => new TaskListResource($taskList),
         ]);
@@ -53,6 +55,8 @@ class TaskListController extends Controller
 
     public function tasks(TaskList $taskList)
     {
+        abort_unless($taskList->user_id === Auth::id(), 403);
+
         return $this->success([
             'tasks' => TaskResource::collection($taskList->tasks),
         ]);
@@ -60,6 +64,8 @@ class TaskListController extends Controller
 
     public function addTask(Request $request, TaskList $taskList)
     {
+        abort_unless($taskList->user_id === Auth::id(), 403);
+
         $request->validate([
             'title' => ['required', 'string'],
             'details' => ['nullable', 'string'],
@@ -83,6 +89,8 @@ class TaskListController extends Controller
      */
     public function update(Request $request, TaskList $taskList)
     {
+        abort_unless($taskList->user_id === Auth::id(), 403);
+
         $request->validate([
             'title' => ['required', 'string']
         ]);
@@ -100,6 +108,8 @@ class TaskListController extends Controller
      */
     public function destroy(TaskList $taskList)
     {
+        abort_unless($taskList->user_id === Auth::id(), 403);
+
         $taskList->delete();
 
         return $this->success();
